requirements.txt
django
whitenoise
modal-client
psycopg2-binary